Stop mineshaft from using 1k power when there are no layers dug. Allow lower amount of electricity to be sent to layers. Lower default amount used. Begin adding dev reset function
Power will always start at 200w base, then that will increase as soon as a layer is dug.
Players can chose to cut power to the layer and the base power consumption will be 200.

Players can now send less than 1k electricity to lower layers if they choose, the default being 100w.

Now the resting electicity usage of the mine shaft is 200w while not sending electricity, minimum 300w while sending electricity with no maximum.

It might throw a div/0 error _if the player sets the output to 100 and has 101 powered mine layers_ but that's a real edge case.

I began to add a gizmo in dev mode that will allow the states of shafts and layers to reset, specifically to correct old bugs in my main save that existed before I started modding it.