Add test dev button. Fix wonky cast that  caused exception.
I did a bit of mucking about trying to fix a `System.InvalidCastException` I've had on my main save for the longest time that occurred when selecting a layer.
It turned out to be a weird cast at line 192 that was fixed by simplifying the code.

While trying to fix that problem, not realizing exactly what it was, I added a gizmo to the mineshaft which only appears in devmode that if pressed will reset underground lifts and reconnect them with an existing map parent.
This isn't useful except to fix errors, maybe if someone had 2 mineshafts but wanted to reduce it to 1 and had issues, this button might help.