Fix initializing variables. Fixed extra power being consumed when no layers are active
Fixed inconsistencies in variables when a save with an older version updates to this one.

Added a check so that the player can't change extra power when no layers are active (because you can't see the amounts in the ui in that state) and stopped the mining shaft from consuming more than idle power when no layers are active.