Tweak variables and add checks for power
Attempted to ensure that if the mod is updated in a save that the new variables will play nice and never go into a weird inconsistent states.
ActiveLayers refuses to initialize to -1 for some reason so I took an alternative approach using a bool.

Added warning option to DeepRim verbose logs.