Fixed OOB and lift use while not powered. Fixed send buttons visible when no layer targeted
The mining shaft's "Bring Up" widget caused items to be sent out of bounds, causing them to despawn and throw an exception.
The code responsible for that didn't really need to exist so I nuked it and instead just called the "Bring Up" method of the targeted lift.

Underground lifts will now refuse to move when not powered, like the main lift. Pawns can now be trapped underground if the power goes out! Oh no!

Hid the send buttons when a new layer is the selected target.