Added the naming stuff to item transfer ui
I forgot to do the same thing to the item transfer level button and InspectString as I did to the level select one. Now they're the same and transfer target uses the names too.

Also fixed that when the parent drill is set to drill to a new layer, the underground lift would say "Target: Layer at depth 00m"
Now it says the same thing as the parent drill, "target new layer at depth"

Target layer readout will say "This layer" instead of "Layer at depth xx (name)" if the layer targeted is the one being viewed (for underground lifts)

In the layer select widget, the current layer will be labeled "##m This layer" instead of the usual label so it stands out. The name of the layer won't be attached but it is visible in the bottom-left box via the inspectString.